MCHECKSTYLE-211 add feature to enable 'inline configuration'
Enable a way to configure the checkstyle rules directly in the pom.
This is an easy way to get rid of needing an own parent-pom structure
for not introducing a build cycle because of the otherwise needed
separate checkstyle-rules project.
Txs to ASF committer rmannibucau for the initial baseline patch.
I reworked it to not need any CDATA section.  


git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@1572667 13f79535-47bb-0310-9956-ffa450edef68